refactor(discord): split voice subsystem by lifecycle owner
