+++ extensions/discord/src/voice/realtime-session.runtime.ts
@@ -1,0 +1,624 @@
+import type { DiscordAccountConfig, OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
+import {
+  buildRealtimeVoiceAgentConsultPolicyInstructions,
+  createRealtimeVoiceSessionHarness,
+  matchRealtimeVoiceConsultQuestions,
+  REALTIME_VOICE_AGENT_CONTROL_TOOL,
+  REALTIME_VOICE_AGENT_CONSULT_TOOL_NAME,
+  REALTIME_VOICE_AUDIO_FORMAT_PCM16_24KHZ,
+  resolveConfiguredRealtimeVoiceProvider,
+  resolveRealtimeVoiceAgentConsultToolPolicy,
+  resolveRealtimeVoiceAgentConsultTools,
+  resolveRealtimeVoiceAgentConsultToolsAllow,
+  type RealtimeVoiceAgentConsultToolPolicy,
+  type RealtimeVoiceBridgeEvent,
+  type RealtimeVoiceBridgeSession,
+  type RealtimeVoiceProviderConfig,
+  type RealtimeVoiceSessionHarness,
+} from "openclaw/plugin-sdk/realtime-voice";
+import { createSubsystemLogger } from "openclaw/plugin-sdk/runtime-env";
+import { formatErrorMessage } from "openclaw/plugin-sdk/ssrf-runtime";
+import { asBoolean } from "openclaw/plugin-sdk/string-coerce-runtime";
+import {
+  isDiscordRealtimeWakeNameRequired,
+  resolveDiscordRealtimeWakeNamePolicy,
+  resolveDiscordRealtimeWakeNames,
+  type DiscordRealtimeWakeNamePolicy,
+} from "./activation.js";
+import { formatVoiceLogPreview } from "./log-preview.js";
+import { DiscordRealtimeConsults, type AgentProxyConsultState } from "./realtime-consults.js";
+import {
+  DiscordRealtimePlayback,
+  resolveDiscordRealtimeBargeIn,
+  resolveDiscordRealtimeMinBargeInAudioEndMs,
+} from "./realtime-playback.js";
+import { DiscordRealtimeTurns } from "./realtime-turns.js";
+import {
+  logVoiceVerbose,
+  type DiscordVoiceMode,
+  type VoiceRealtimeAgentTurnParams,
+  type VoiceRealtimeSession,
+  type VoiceRealtimeSpeakerContext,
+  type VoiceRealtimeSpeakerTurn,
+  type VoiceSessionEntry,
+} from "./session.js";
+
+const logger = createSubsystemLogger("discord/voice");
+const DISCORD_REALTIME_DUPLICATE_ERROR_SUPPRESS_MS = 60_000;
+const discordRealtimeTalkPayload = () => ({});
+const DISCORD_REALTIME_VERBOSE_OMITTED_EVENTS = new Set([
+  "conversation.output_audio.delta",
+  "input_audio_buffer.append",
+  "response.audio.delta",
+  "response.output_audio.delta",
+]);
+
+type DiscordRealtimeVoiceConfig = NonNullable<DiscordAccountConfig["voice"]>["realtime"];
+
+function resolveDiscordRealtimeVoiceAgentConsultTools(policy: RealtimeVoiceAgentConsultToolPolicy) {
+  const tools = resolveRealtimeVoiceAgentConsultTools(policy);
+  if (
+    policy !== "none" &&
+    !tools.some((tool) => tool.name === REALTIME_VOICE_AGENT_CONTROL_TOOL.name)
+  ) {
+    return [...tools, REALTIME_VOICE_AGENT_CONTROL_TOOL];
+  }
+  return tools;
+}
+
+function formatRealtimeInterruptionLog(event: RealtimeVoiceBridgeEvent): string | undefined {
+  const detail = event.detail ? ` ${event.detail}` : "";
+  if (event.direction === "client") {
+    if (event.type === "response.cancel") {
+      return `discord voice: realtime model interrupt requested ${event.direction}:${event.type}${detail}`;
+    }
+    if (event.type === "conversation.item.truncate.skipped") {
+      return `discord voice: realtime model interrupt ignored ${event.direction}:${event.type}${detail}`;
+    }
+    if (event.type === "conversation.item.truncate") {
+      return `discord voice: realtime model audio truncated ${event.direction}:${event.type}${detail}`;
+    }
+  }
+  if (event.direction === "server") {
+    if (event.type === "response.cancelled") {
+      return `discord voice: realtime model interrupt confirmed ${event.direction}:${event.type}${detail}`;
+    }
+    if (
+      event.type === "error" &&
+      event.detail === "Cancellation failed: no active response found"
+    ) {
+      return `discord voice: realtime model interrupt raced ${event.direction}:${event.type}${detail}`;
+    }
+  }
+  return undefined;
+}
+
+function formatRealtimeLifecycleLog(event: RealtimeVoiceBridgeEvent): string | undefined {
+  if (!event.type.startsWith("session.")) {
+    return undefined;
+  }
+  const detail = event.detail ? ` ${event.detail}` : "";
+  return `discord voice: realtime lifecycle ${event.direction}:${event.type}${detail}`;
+}
+
+function shouldLogRealtimeVerboseEvent(event: RealtimeVoiceBridgeEvent): boolean {
+  return !DISCORD_REALTIME_VERBOSE_OMITTED_EVENTS.has(event.type);
+}
+
+function readProviderConfigString(
+  config: RealtimeVoiceProviderConfig,
+  key: string,
+): string | undefined {
+  const value = config[key];
+  return typeof value === "string" && value.trim() ? value.trim() : undefined;
+}
+
+function isDiscordAgentProxyVoiceMode(mode: DiscordVoiceMode): boolean {
+  return mode === "agent-proxy";
+}
+
+// Follow-up: replace this string protocol with a typed realtime-voice talk outcome.
+// Keep the exact-speech scraper byte-stable until that owner-reviewed task.
+function buildDiscordSpeakExactUserMessage(text: string): string {
+  return [
+    "Internal OpenClaw voice playback result.",
+    "Do not call openclaw_agent_consult or any other tool for this message.",
+    "Speak this exact OpenClaw answer to the Discord voice channel, without adding, removing, or rephrasing words.",
+    `Answer: ${JSON.stringify(text)}`,
+  ].join("\n");
+}
+
+function isEscapedQuote(text: string, quoteIndex: number): boolean {
+  let backslashes = 0;
+  for (let index = quoteIndex - 1; index >= 0 && text[index] === "\\"; index -= 1) {
+    backslashes += 1;
+  }
+  return backslashes % 2 === 1;
+}
+
+function readJsonStringAfterLabel(text: string, label: string): string | undefined {
+  const labelIndex = text.indexOf(label);
+  if (labelIndex < 0) {
+    return undefined;
+  }
+  const quoteIndex = text.indexOf('"', labelIndex + label.length);
+  if (quoteIndex < 0) {
+    return undefined;
+  }
+  for (let index = quoteIndex + 1; index < text.length; index += 1) {
+    if (text[index] !== '"' || isEscapedQuote(text, index)) {
+      continue;
+    }
+    try {
+      const parsed: unknown = JSON.parse(text.slice(quoteIndex, index + 1));
+      return typeof parsed === "string" ? parsed : undefined;
+    } catch {
+      return undefined;
+    }
+  }
+  return undefined;
+}
+
+function collectRealtimeConsultArgStrings(args: unknown): string[] {
+  if (!args || typeof args !== "object") {
+    return typeof args === "string" ? [args] : [];
+  }
+  const values: string[] = [];
+  for (const key of ["question", "prompt", "query", "task", "context", "responseStyle"]) {
+    const value = (args as Record<string, unknown>)[key];
+    if (typeof value === "string") {
+      values.push(value);
+    }
+  }
+  return values;
+}
+
+function extractDiscordExactSpeechConsultText(args: unknown): string | undefined {
+  const message = collectRealtimeConsultArgStrings(args).join("\n");
+  if (
+    !message.includes("Speak this exact OpenClaw answer") &&
+    !message.includes("Speak the provided exact answer verbatim")
+  ) {
+    return undefined;
+  }
+  return (
+    readJsonStringAfterLabel(message, "Answer:") ??
+    readJsonStringAfterLabel(message, "Provided answer text:")
+  );
+}
+
+export class DiscordRealtimeVoiceSession implements VoiceRealtimeSession {
+  private bridge: RealtimeVoiceBridgeSession | null = null;
+  private readonly harness: RealtimeVoiceSessionHarness<AgentProxyConsultState>;
+  private readonly playback: DiscordRealtimePlayback<AgentProxyConsultState>;
+  private readonly turns: DiscordRealtimeTurns;
+  private readonly consults: DiscordRealtimeConsults;
+  private stopped = false;
+  private consultToolPolicy: RealtimeVoiceAgentConsultToolPolicy = "safe-read-only";
+  private consultToolsAllow: string[] | undefined;
+  private consultPolicy: "auto" | "always" = "auto";
+  private wakeNamePolicy: DiscordRealtimeWakeNamePolicy = "never";
+  private wakeNames: string[] = [];
+  private realtimeProviderId: string | undefined;
+  private bridgeReady = false;
+  private providerGenerationObserved = false;
+  private providerContinuityEpoch = 0;
+  private lastRealtimeError:
+    | { message: string; suppressed: number; lastLoggedAt: number }
+    | undefined;
+
+  constructor(
+    private readonly params: {
+      cfg: OpenClawConfig;
+      discordConfig: DiscordAccountConfig;
+      entry: VoiceSessionEntry;
+      mode: Exclude<DiscordVoiceMode, "stt-tts">;
+      bootstrapContextInstructions?: string;
+      getHumanParticipantCount?: () => number;
+      onTerminalError: (error: Error) => void;
+      runAgentTurn: (params: VoiceRealtimeAgentTurnParams) => Promise<string>;
+    },
+  ) {
+    this.harness = createRealtimeVoiceSessionHarness<AgentProxyConsultState>({
+      talk: {
+        sessionId: `discord:${this.params.entry.voiceSessionKey}:realtime`,
+        mode: "realtime",
+        transport: "gateway-relay",
+        brain: "agent-consult",
+      },
+      talkPayloads: {
+        turnStarted: discordRealtimeTalkPayload,
+        turnEnded: discordRealtimeTalkPayload,
+        inputAudioDelta: discordRealtimeTalkPayload,
+        outputAudioStarted: discordRealtimeTalkPayload,
+        outputAudioDelta: discordRealtimeTalkPayload,
+        outputAudioDone: discordRealtimeTalkPayload,
+      },
+      forcedConsults: {
+        limit: 16,
+        nativeDedupeMs: 15_000,
+        questionsMatch: matchRealtimeVoiceConsultQuestions,
+      },
+    });
+    this.playback = new DiscordRealtimePlayback({
+      bridge: () => this.bridge,
+      bridgeReady: () => this.bridgeReady,
+      buildSpeakExactMessage: buildDiscordSpeakExactUserMessage,
+      entry: this.params.entry,
+      harness: this.harness,
+      markProviderGenerationObserved: () => this.markProviderGenerationObserved(),
+      mode: this.params.mode,
+      onTerminalError: this.params.onTerminalError,
+      providerId: () => this.realtimeProviderId,
+      realtimeConfig: () => this.realtimeConfig,
+      stopTerminally: () => {
+        this.stopped = true;
+        this.bridgeReady = false;
+        this.consults.close();
+      },
+      stopped: () => this.stopped,
+      wakeNameRequired: () => this.isWakeNameRequired(),
+    });
+    this.turns = new DiscordRealtimeTurns({
+      bridge: () => this.bridge,
+      entry: this.params.entry,
+      getHumanParticipantCount: () => this.humanParticipantCount(),
+      onAcceptedTranscript: (text, context, providerEpoch) =>
+        this.consults.handleAcceptedTranscript(text, context, providerEpoch),
+      playback: this.playback,
+      providerEpoch: () => this.providerContinuityEpoch,
+      providerId: () => this.realtimeProviderId,
+      realtimeConfig: () => this.realtimeConfig,
+      recordInputAudio: (audio) => this.harness.recordInputAudio(audio),
+      stopped: () => this.stopped,
+      wakeNamePolicy: () => this.wakeNamePolicy,
+      wakeNames: () => this.wakeNames,
+    });
+    this.consults = new DiscordRealtimeConsults({
+      consultPolicy: () => this.consultPolicy,
+      consultToolPolicy: () => this.consultToolPolicy,
+      consultToolsAllow: () => this.consultToolsAllow,
+      debounceMs: () => this.realtimeConfig?.debounceMs,
+      entry: this.params.entry,
+      extractExactSpeech: extractDiscordExactSpeechConsultText,
+      harness: this.harness,
+      isAgentProxy: isDiscordAgentProxyVoiceMode(this.params.mode),
+      isWakeNameRequired: () => this.isWakeNameRequired(),
+      playback: this.playback,
+      providerEpoch: () => this.providerContinuityEpoch,
+      runAgentTurn: this.params.runAgentTurn,
+      stopped: () => this.stopped,
+      turns: this.turns,
+      usesRealtimeAgentHandoff: () =>
+        this.params.mode === "bidi" || this.consultToolPolicy !== "none",
+      wakeNamePolicy: () => this.wakeNamePolicy,
+    });
+  }
+
+  async connect(): Promise<void> {
+    const resolved = resolveConfiguredRealtimeVoiceProvider({
+      configuredProviderId: this.realtimeConfig?.provider,
+      providerConfigs: buildProviderConfigs(this.realtimeConfig),
+      providerConfigOverrides: buildProviderConfigOverrides(this.realtimeConfig),
+      cfg: this.params.cfg,
+      defaultModel: this.realtimeConfig?.model,
+      noRegisteredProviderMessage: "No configured realtime voice provider registered",
+    });
+    this.realtimeProviderId = resolved.provider.id;
+    const isAgentProxy = isDiscordAgentProxyVoiceMode(this.params.mode);
+    // Follow-up: move generic agent-proxy instructions and consult/tool policy to realtime-voice.
+    // Keep this policy block byte-stable until that owner-reviewed task.
+    const defaultToolPolicy: RealtimeVoiceAgentConsultToolPolicy = isAgentProxy
+      ? "owner"
+      : "safe-read-only";
+    const toolPolicy = resolveRealtimeVoiceAgentConsultToolPolicy(
+      this.realtimeConfig?.toolPolicy,
+      defaultToolPolicy,
+    );
+    this.consultToolPolicy = toolPolicy;
+    this.consultToolsAllow = resolveRealtimeVoiceAgentConsultToolsAllow(toolPolicy);
+    const consultPolicy = this.realtimeConfig?.consultPolicy ?? (isAgentProxy ? "always" : "auto");
+    this.consultPolicy = consultPolicy;
+    this.wakeNamePolicy = resolveDiscordRealtimeWakeNamePolicy({
+      isAgentProxy,
+      providerId: resolved.provider.id,
+      requireWakeName: this.realtimeConfig?.requireWakeName,
+    });
+    this.wakeNames =
+      this.wakeNamePolicy !== "never"
+        ? resolveDiscordRealtimeWakeNames({
+            config: this.realtimeConfig,
+            cfg: this.params.cfg,
+            agentId: this.params.entry.route.agentId,
+          })
+        : [];
+    const usesRealtimeAgentHandoff = this.params.mode === "bidi" || toolPolicy !== "none";
+    const autoRespondToAudio =
+      this.wakeNamePolicy === "never" && (!isAgentProxy || consultPolicy !== "always");
+    const interruptResponseOnInputAudio =
+      this.wakeNamePolicy === "never" &&
+      resolveDiscordRealtimeInterruptResponseOnInputAudio({
+        realtimeConfig: this.realtimeConfig,
+        providerId: resolved.provider.id,
+      });
+    const instructions = buildDiscordRealtimeInstructions({
+      mode: this.params.mode,
+      instructions: this.realtimeConfig?.instructions,
+      bootstrapContextInstructions: this.params.bootstrapContextInstructions,
+      toolPolicy,
+      consultPolicy,
+    });
+    this.bridge = this.harness.createBridge({
+      provider: resolved.provider,
+      cfg: this.params.cfg,
+      providerConfig: resolved.providerConfig,
+      audioFormat: REALTIME_VOICE_AUDIO_FORMAT_PCM16_24KHZ,
+      instructions,
+      autoRespondToAudio,
+      interruptResponseOnInputAudio,
+      markStrategy: "ack-immediately",
+      tools: usesRealtimeAgentHandoff
+        ? resolveDiscordRealtimeVoiceAgentConsultTools(toolPolicy)
+        : [],
+      audioSink: {
+        isOpen: () => !this.stopped,
+        sendAudio: (audio) => this.playback.sendOutputAudio(audio),
+        clearAudio: () => {
+          this.markProviderGenerationObserved();
+          this.harness.flushOutput(() => this.playback.clearOutputAudio("provider-clear-audio"));
+        },
+      },
+      onTranscript: (role, text, isFinal) => {
+        this.markProviderGenerationObserved();
+        if (isFinal && text.trim()) {
+          logger.info(
+            `discord voice: realtime ${role} transcript (${text.length} chars): ${formatVoiceLogPreview(text)}`,
+          );
+        }
+        if (isFinal && role === "assistant") {
+          this.playback.suppressDuplicateControlSpeech(text);
+        }
+        if (role !== "user") {
+          return;
+        }
+        if (!isFinal) {
+          this.turns.handlePartialUserTranscript(text);
+          return;
+        }
+        void this.turns.handleFinalUserTranscript(text);
+      },
+      onToolCall: (event, session) => {
+        this.markProviderGenerationObserved();
+        return this.consults.handleToolCall(event, session);
+      },
+      onReady: () => {
+        this.markProviderGenerationObserved();
+        this.bridgeReady = true;
+        this.playback.drainQueuedExactSpeechMessages("provider-ready");
+      },
+      onEvent: (event) => this.handleBridgeEvent(event),
+      onResponseDone: (outcome) => {
+        this.markProviderGenerationObserved();
+        this.playback.handleResponseDone(outcome);
+        if (outcome.status === "cancelled") {
+          logger.info(
+            `discord voice: realtime model interrupt confirmed server:response.done status=cancelled${outcome.reason ? ` reason=${outcome.reason}` : ""}`,
+          );
+        } else if (outcome.status === "failed" || outcome.status === "incomplete") {
+          this.logRealtimeError(outcome.message);
+        }
+      },
+      onError: (error) => this.logRealtimeError(formatErrorMessage(error)),
+      onClose: (reason) => {
+        this.flushSuppressedRealtimeErrors();
+        logVoiceVerbose(`realtime closed: ${reason}`);
+      },
+    });
+    const resolvedModel =
+      readProviderConfigString(resolved.providerConfig, "model") ?? resolved.provider.defaultModel;
+    const resolvedVoice = readProviderConfigString(resolved.providerConfig, "voice");
+    const humanParticipantCount = this.humanParticipantCount();
+    logger.info(
+      `discord voice: realtime bridge starting mode=${this.params.mode} provider=${resolved.provider.id} model=${resolvedModel ?? "default"} voice=${resolvedVoice ?? "default"} consultPolicy=${consultPolicy} toolPolicy=${toolPolicy} autoRespond=${autoRespondToAudio} wakeNamePolicy=${this.wakeNamePolicy} requireWakeName=${this.isWakeNameRequired(humanParticipantCount)} humanParticipants=${humanParticipantCount} wakeNames=${this.wakeNames.join(",") || "none"} interruptResponse=${interruptResponseOnInputAudio} bargeIn=${resolveDiscordRealtimeBargeIn({ realtimeConfig: this.realtimeConfig, providerId: resolved.provider.id })} minBargeInAudioEndMs=${resolveDiscordRealtimeMinBargeInAudioEndMs(this.realtimeConfig)}`,
+    );
+    this.playback.attachPlayer();
+    await this.bridge.connect();
+    this.markProviderGenerationObserved();
+    this.bridgeReady = true;
+    this.playback.drainQueuedExactSpeechMessages("provider-connected");
+    logger.info(
+      `discord voice: realtime bridge ready mode=${this.params.mode} provider=${resolved.provider.id} model=${resolvedModel ?? "default"} voice=${resolvedVoice ?? "default"}`,
+    );
+  }
+
+  close(): void {
+    this.stopped = true;
+    this.bridgeReady = false;
+    this.providerContinuityEpoch += 1;
+    this.flushSuppressedRealtimeErrors();
+    this.consults.close();
+    this.harness.close();
+    this.turns.clear();
+    this.playback.close();
+    this.bridge?.close();
+    this.bridge = null;
+    this.realtimeProviderId = undefined;
+  }
+
+  beginSpeakerTurn(context: VoiceRealtimeSpeakerContext, userId: string): VoiceRealtimeSpeakerTurn {
+    return this.turns.beginSpeakerTurn(context, userId);
+  }
+
+  handleBargeIn(reason = "barge-in"): void {
+    this.playback.handleBargeIn(reason);
+  }
+
+  isBargeInEnabled(): boolean {
+    if (this.isWakeNameRequired()) {
+      return false;
+    }
+    return this.playback.isBargeInEnabled();
+  }
+
+  private get outputStream() {
+    return this.playback.currentOutputStream();
+  }
+
+  private enqueueExactSpeechMessage(text: string): void {
+    this.playback.enqueueExactSpeechMessage(text);
+  }
+
+  private get realtimeConfig(): DiscordRealtimeVoiceConfig {
+    return this.params.discordConfig.voice?.realtime;
+  }
+
+  private humanParticipantCount(): number {
+    return this.params.getHumanParticipantCount?.() ?? 0;
+  }
+
+  private isWakeNameRequired(humanParticipantCount = this.humanParticipantCount()): boolean {
+    return isDiscordRealtimeWakeNameRequired(this.wakeNamePolicy, humanParticipantCount);
+  }
+
+  private handleBridgeEvent(event: RealtimeVoiceBridgeEvent): void {
+    if (!(event.direction === "client" && event.type === "session.continuity.reset")) {
+      this.markProviderGenerationObserved();
+    }
+    const detail = event.detail ? ` ${event.detail}` : "";
+    if (event.direction === "client" && event.type === "session.continuity.reset") {
+      this.resetProviderContinuity(event.type);
+    }
+    if (event.direction === "server" && event.type === "input_audio_buffer.speech_started") {
+      this.turns.resetPartialWakeNameTracking();
+    }
+    if (shouldLogRealtimeVerboseEvent(event)) {
+      logVoiceVerbose(`realtime ${event.direction}:${event.type}${detail}`);
+    }
+    this.playback.handleProviderEvent(event);
+    const interruptionLog = formatRealtimeInterruptionLog(event);
+    if (interruptionLog) {
+      logger.info(interruptionLog);
+    }
+    const lifecycleLog = formatRealtimeLifecycleLog(event);
+    if (lifecycleLog) {
+      logger.info(lifecycleLog);
+    }
+  }
+
+  private markProviderGenerationObserved(): void {
+    this.providerGenerationObserved = true;
+  }
+
+  private resetProviderContinuity(reason: string): void {
+    if (!this.providerGenerationObserved) {
+      return;
+    }
+    this.providerGenerationObserved = false;
+    this.bridgeReady = false;
+    this.providerContinuityEpoch += 1;
+    this.consults.resetProviderContinuity();
+    this.turns.resetProviderContinuity();
+    this.playback.resetProviderContinuity(reason);
+  }
+
+  private logRealtimeError(message: string): void {
+    const now = Date.now();
+    if (
+      this.lastRealtimeError?.message === message &&
+      now - this.lastRealtimeError.lastLoggedAt < DISCORD_REALTIME_DUPLICATE_ERROR_SUPPRESS_MS
+    ) {
+      this.lastRealtimeError.suppressed += 1;
+      return;
+    }
+    this.flushSuppressedRealtimeErrors();
+    this.lastRealtimeError = { message, suppressed: 0, lastLoggedAt: now };
+    logger.warn(`discord voice: realtime error: ${message}`);
+  }
+
+  private flushSuppressedRealtimeErrors(): void {
+    if (!this.lastRealtimeError || this.lastRealtimeError.suppressed === 0) {
+      return;
+    }
+    logger.warn(
+      `discord voice: suppressed ${this.lastRealtimeError.suppressed} duplicate realtime errors: ${this.lastRealtimeError.message}`,
+    );
+    this.lastRealtimeError.suppressed = 0;
+  }
+}
+
+function buildProviderConfigs(
+  realtimeConfig: DiscordRealtimeVoiceConfig,
+): Record<string, RealtimeVoiceProviderConfig | undefined> | undefined {
+  const configs = realtimeConfig?.providers;
+  return configs && Object.keys(configs).length > 0 ? { ...configs } : undefined;
+}
+
+function buildProviderConfigOverrides(
+  realtimeConfig: DiscordRealtimeVoiceConfig,
+): RealtimeVoiceProviderConfig | undefined {
+  const overrides = {
+    ...(realtimeConfig?.model ? { model: realtimeConfig.model } : {}),
+    ...(realtimeConfig?.speakerVoice
+      ? { voice: realtimeConfig.speakerVoice }
+      : realtimeConfig?.speakerVoiceId
+        ? { voice: realtimeConfig.speakerVoiceId }
+        : {}),
+    ...(typeof realtimeConfig?.minBargeInAudioEndMs === "number"
+      ? { minBargeInAudioEndMs: realtimeConfig.minBargeInAudioEndMs }
+      : {}),
+  };
+  return Object.keys(overrides).length > 0 ? overrides : undefined;
+}
+
+function resolveDiscordRealtimeInterruptResponseOnInputAudio(params: {
+  realtimeConfig: DiscordRealtimeVoiceConfig;
+  providerId: string;
+}): boolean {
+  const value =
+    params.realtimeConfig?.providers?.[params.providerId]?.interruptResponseOnInputAudio;
+  return asBoolean(value) ?? true;
+}
+
+function buildDiscordRealtimeInstructions(params: {
+  mode: Exclude<DiscordVoiceMode, "stt-tts">;
+  instructions?: string;
+  bootstrapContextInstructions?: string;
+  toolPolicy: RealtimeVoiceAgentConsultToolPolicy;
+  consultPolicy: "auto" | "always";
+}): string {
+  const base =
+    params.instructions ??
+    [
+      "You are OpenClaw's Discord voice interface.",
+      "Keep spoken replies concise, natural, and suitable for a live Discord voice channel.",
+    ].join("\n");
+  if (isDiscordAgentProxyVoiceMode(params.mode)) {
+    return [
+      base,
+      params.bootstrapContextInstructions?.trim(),
+      "Mode: OpenClaw agent proxy.",
+      "You are the realtime voice surface for the same OpenClaw agent the user can message directly.",
+      "Do not mention a backend, supervisor, helper, or separate system. Present the result as your own work.",
+      "Delegate substantive requests, actions, tool work, current facts, memory, workspace context, and user-specific context with openclaw_agent_consult.",
+      "Do not block, refuse, or downscope at the voice layer. Delegate to OpenClaw and treat its result as authoritative.",
+      "Answer directly only for greetings, acknowledgements, brief latency tests, or filler while waiting.",
+      'While waiting for OpenClaw data or tool results, use at most one short natural backchannel such as "yeah", "mm-hmm", "got it", or "one sec"; vary it and do not treat it as the final answer.',
+      "When OpenClaw sends an internal exact answer to speak, do not call tools. Say only that answer.",
+      buildRealtimeVoiceAgentConsultPolicyInstructions({
+        toolPolicy: params.toolPolicy,
+        consultPolicy: params.consultPolicy,
+      }),
+    ].join("\n\n");
+  }
+  return [
+    base,
+    params.bootstrapContextInstructions?.trim(),
+    'While waiting for OpenClaw data or tool results, use at most one short natural backchannel such as "yeah", "mm-hmm", "got it", or "one sec"; vary it and do not treat it as the final answer.',
+    buildRealtimeVoiceAgentConsultPolicyInstructions({
+      toolPolicy: params.toolPolicy,
+      consultPolicy: params.consultPolicy,
+    }),
+  ]
+    .filter(Boolean)
+    .join("\n\n");
+}
